Add a configuration special to China
Their NTP server requirements differ, we leave the RHEL pool as
the default but change the backup servers to ones which we are
certain will work in the region.

Thanks to Aefer for coming up with the idea of checking the GSM
country code to detect whether a user was in China or not, this
complimentes PR #4.

Co-authored-by: Aefer <24859058+Aefer@users.noreply.github.com>
Signed-off-by: Laster K. (lazerl0rd) <officiallazerl0rd@gmail.com>